git-svn-id: https://opengl-scut.googlecode.com/svn/trunk@439 81c3a017-244d-672b-1dc6-c8d6962b81c2

+++ Code/Scar/Scar/Toolkit.h
@@ -19,7 +19,7 @@ struct Node2DInfo
 {
 	s32 width;
 	s32 height;
-	position2d<s32> pos;
+	position2d<f32> pos;
 };
